requirements.txt
requests
openai
python-dotenv
markdownify